Refactor as leaf_filter_fxn
