+++ timemachine/fe/mcgregor.py
@@ -3,18 +3,11 @@ import copy
 import time
 import warnings
 from dataclasses import dataclass
-from typing import Callable, List, Optional, Sequence, Set, Tuple
+from typing import Callable, List, Mapping, Optional, Sequence, Set, Tuple, TypeAlias
 
 import networkx as nx
 import numpy as np
 from numpy.typing import NDArray
-
-from timemachine.fe.dummy import (
-    canonicalize_bond,
-    compute_disabled_bonds_in_core,
-    compute_disabled_bonds_in_dga,
-    generate_dummy_group_assignments,
-)
 
 
 def _arcs_left(marcs):
@@ -299,6 +292,17 @@ def perm_to_core(perm: Sequence[int]) -> NDArray:
     return core_array
 
 
+LeafFilterFxn: TypeAlias = Callable[
+    [
+        nx.Graph,
+        nx.Graph,
+        Sequence[int] | Mapping[int, int],
+        Sequence[int] | Mapping[int, int],
+    ],
+    bool,
+]
+
+
 def mcs(
     n_a,
     n_b,
@@ -312,8 +316,8 @@ def mcs(
     min_connected_component_size: int,
     min_threshold,
     initial_mapping,
-    disallow_chiral_conversion: bool,
     filter_fxn: Callable[[Sequence[int]], bool] = lambda core: True,
+    leaf_filter_fxn: LeafFilterFxn = lambda *_: True,
 ) -> Tuple[List[NDArray], List[NDArray], MCSDiagnostics]:
     assert n_a <= n_b
     assert max_connected_components is None or max_connected_components > 0, "Must have max_connected_components > 0"
@@ -364,8 +368,8 @@ def mcs(
             enforce_core_core,
             max_connected_components,
             min_connected_component_size,
-            disallow_chiral_conversion,
             filter_fxn,
+            leaf_filter_fxn,
         )
 
         total_nodes_visited += mcs_result.nodes_visited
@@ -421,40 +425,6 @@ def atom_map_pop(map_1_to_2, map_2_to_1, idx, jdx):
     map_2_to_1[jdx] = UNMAPPED
 
 
-def get_invalid_chiral_conversion(
-    bond_graph: nx.Graph, disabled_bonds: Set[Tuple[int, int]]
-) -> Optional[Tuple[int, List[int]]]:
-    for node in bond_graph.nodes():
-        nbs = list(nx.neighbors(bond_graph, node))
-        # TODO: handle X3 chiral centers
-        if len(nbs) == 4:
-            disabled_bonds_count = sum(1 for nb in nbs if canonicalize_bond((node, nb)) in disabled_bonds)
-            if disabled_bonds_count > 1:
-                return node, nbs
-    return None
-
-
-def _graph_fails_chiral_assertion(
-    bond_graph: nx.Graph, core_nodes: Set[int], core_disabled_bonds: Set[Tuple[int, int]]
-) -> bool:
-    """A graph fails the chiral assertion if every choice of dummy group assignments fails.
-
-    A particular dummy group assignment fails if there exists a 4-connected atom with more than a single broken bond.
-    """
-    if len(core_nodes) == bond_graph.number_of_nodes() and get_invalid_chiral_conversion(
-        bond_graph, core_disabled_bonds
-    ):
-        return True
-
-    for dga in generate_dummy_group_assignments(bond_graph, core_nodes):
-        dga_disabled_bonds = compute_disabled_bonds_in_dga(bond_graph, core_nodes, dga)
-        disabled_bonds = dga_disabled_bonds.union(core_disabled_bonds)
-        if get_invalid_chiral_conversion(bond_graph, disabled_bonds):
-            return True
-
-    return False
-
-
 def recursion(
     g1: Graph,
     g2: Graph,
@@ -470,8 +440,8 @@ def recursion(
     enforce_core_core,
     max_connected_components: Optional[int],
     min_connected_component_size: int,
-    disallow_chiral_conversion: bool,
     filter_fxn,
+    leaf_filter_fxn,
 ):
     if mcs_result.nodes_visited > max_visits:
         mcs_result.timed_out = True
@@ -513,28 +483,8 @@ def recursion(
     # leaf-node, every atom has been mapped
     if layer == n_a:
         if num_edges == threshold:
-            if disallow_chiral_conversion:
-                # # chiral assertion check on leaf nodes
-                g1_mapped_nodes = {a1 for a1, a2 in enumerate(atom_map_1_to_2) if a2 != UNMAPPED}
-                g1_core_disabled_bonds = compute_disabled_bonds_in_core(
-                    g1.nxg, g2.nxg, g1_mapped_nodes, atom_map_1_to_2
-                )
-                if enforce_core_core:
-                    assert len(g1_core_disabled_bonds) == 0
-
-                if _graph_fails_chiral_assertion(g1.nxg, g1_mapped_nodes, g1_core_disabled_bonds):
-                    return
-
-                g2_mapped_nodes = {a2 for a2, a1 in enumerate(atom_map_2_to_1) if a1 != UNMAPPED}
-                g2_core_disabled_bonds = compute_disabled_bonds_in_core(
-                    g2.nxg, g1.nxg, g2_mapped_nodes, atom_map_2_to_1
-                )
-
-                if enforce_core_core:
-                    assert len(g2_core_disabled_bonds) == 0
-
-                if _graph_fails_chiral_assertion(g2.nxg, g2_mapped_nodes, g2_core_disabled_bonds):
-                    return
+            if not leaf_filter_fxn(g1.nxg, g2.nxg, atom_map_1_to_2, atom_map_2_to_1):
+                return
 
             mcs_result.all_maps.append(copy.copy(atom_map_1_to_2))
             mcs_result.all_marcs.append(copy.copy(marcs))
@@ -567,8 +517,8 @@ def recursion(
                     enforce_core_core,
                     max_connected_components,
                     min_connected_component_size,
-                    disallow_chiral_conversion,
                     filter_fxn,
+                    leaf_filter_fxn,
                 )
             atom_map_pop(atom_map_1_to_2, atom_map_2_to_1, layer, jdx)
 
@@ -591,6 +541,6 @@ def recursion(
         enforce_core_core,
         max_connected_components,
         min_connected_component_size,
-        disallow_chiral_conversion,
         filter_fxn,
+        leaf_filter_fxn,
     )
